Still unworking state - Dabsic mods not achieved yet.
Dabsic miss litteral arrays/scopes and the construction that would be nice to have (function that return a scope, for example, presented as a scope):

[Stuff(int x)
  Var = x * 2
  Var2 = x * x
]

{Stuff(real n)
  n * n,
  3 * n,
}

Etc. A way to achieve that would be to automaticaly turn all fields inside to functions with the same
prototype as the main function node and to automatically pass parameters. The result would be copied
from the main function.